requirements.txt
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.9.0 ; python_version >= "3.11" and python_version < "4.0"
beanie==1.29.0 ; python_version >= "3.11" and python_version < "4.0"
click==8.2.1 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
dnspython==2.7.0 ; python_version >= "3.11" and python_version < "4.0"
fastapi==0.115.12 ; python_version >= "3.11" and python_version < "4.0"
h11==0.16.0 ; python_version >= "3.11" and python_version < "4.0"
idna==3.10 ; python_version >= "3.11" and python_version < "4.0"
lazy-model==0.2.0 ; python_version >= "3.11" and python_version < "4.0"
mangum==0.19.0 ; python_version >= "3.11" and python_version < "4.0"
motor==3.7.1 ; python_version >= "3.11" and python_version < "4.0"
pydantic-core==2.33.2 ; python_version >= "3.11" and python_version < "4.0"
pydantic-settings==2.9.1 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.11.5 ; python_version >= "3.11" and python_version < "4.0"
pymongo==4.13.0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.1.0 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
starlette==0.46.2 ; python_version >= "3.11" and python_version < "4.0"
toml==0.10.2 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.14.0 ; python_version >= "3.11" and python_version < "4.0"
typing-inspection==0.4.1 ; python_version >= "3.11" and python_version < "4.0"
uvicorn==0.34.3 ; python_version >= "3.11" and python_version < "4.0"
